powens/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/powens"]

[project]
name = "powens"
authors = [
  { name="Jules Anderson" },
]
description = "Package to talk to the Powens API."
version = "2025.0.0"
requires-python = ">= 3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "requests",
    "python-dateutil"
]

[project.urls]
Homepage = "https://github.com/Azurennn/powens-firefly"
Issues = "https://github.com/Azurennn/powens-firefly/issues"

[project.scripts]
jinks = "jinks_core.__main__:main"
